refactor(gc): colocate repl_test and rename runtime to gc_runtime
Move the REPL integration test out of tests/ into the crate root to match
other colocated *_test.rs files, and rename runtime.rs to gc_runtime.rs.
