+++ gc/lib.rs
@@ -12,8 +12,8 @@ pub mod header;
 pub mod heap;
 pub mod list;
 pub mod malloc;
+pub mod gc_runtime;
 pub mod report;
-pub mod runtime;
 pub mod value;
 pub mod vm;
 
@@ -26,7 +26,7 @@ pub use report::{
     GcStatsBundle, GlobalRoot, HashKeyKind, HeapSnapshot, ObjectDecision, RestorationWitness,
     ScanStats, TrialDecision, TrialDeletionStats, ValueKindCounts, VisitedEdge,
 };
-pub use runtime::{GcObject, GcRuntime, MarkFunc};
+pub use gc_runtime::{GcObject, GcRuntime, MarkFunc};
 pub use value::{
     export_object, import_object, try_export_object, value_to_string, GcClosure, Value, ValueKind,
 };
