+++ gc/list.rs
@@ -1,5 +1,5 @@
 use crate::header::GcId;
-use crate::runtime::GcRuntime;
+use crate::gc_runtime::GcRuntime;
 
 /// Intrusive doubly-linked list head/tail, indexed by `GcId`.
 /// Rust equivalent of QuickJS `list.h` + embedded `list_head`.
